windows/build: fix old libssl copy failure
Msys2 updated libssl and the hardcoded name changed. This causes builds
of Memento to be missing runtime dependencies like libssl and qtsvg when
building on Windows. This fixes the issue by using a * in place of
versioning for libssl. It also adds 'set -e' to the script so that build
pipelines will fail with the build script.