linux/gentoo: add ebuilds for 1.0.0 and live git
Add ebuilds for version 1.0.0 and the live version of Memento. Something
to do is to add OCR as a useflag. The problem is currently OCR is
dubious on Gentoo since most of manga-ocr's depdencies are not in the
Gentoo respository. Although only Python is required at compile time, it
is questionable if OCR should be included at all on Gentoo at this point
in time.