overlay/playeroverlay: show and hide subtitles based on cursor position
Allow the user to set whether or not to show primary and secondary
subtitle based on cursor position. Right now primary subtitles are
hardcoded to show if the cursor is on the bottom 20% of the screen.
Likewise secondary subtitles are hardcoded to show at the top 20% of the
screen. This assuming that subtitles are positioned in these two regions
as is convention. If someone is a freak and likes primary subtitles
located at the top of the screen for some reason, this setting will not
work for them. The only reason I don't have adjustable regions for
showing subtitles is due to the complexity of the settings UI. If I get
around to allowing for config files to be edited, I will absolutely add
adjustable regions.

This is made possible by a new HitTestWidget. This inherit from
a QWidget and overrides the mouseMoveEvent() to watch for mouse
movements. It can contain multiple regions to handle actions for, so
when the cursor interacts with that region it will automatically fire
off the region handler. This makes it easy to add more functionality
based on cursor position in the future.

All of this functionality is disabled by default.

Closes #119