widgets/definition/termwidget: fix false positives when reading as exp
If a term with the same reading as another was added using reading as
expression, all terms with the same reading would be regarded as
duplicates. This fixes the issue by treating expression and reading as
seperately addable. This means there can now be up to 5 buttons visible
at a time in a search. Fun.

Closes #151