src/player/mpvadapter: handle key to text conversion more robustly
The current code fails to work properly on macOS because
QKeyEvent::text() doesn't work as expected when modifier keys are held.
This circumvents all that nonsense to just handle converting the keycode
directly which works in most cases. Only obvious case it fails in is
when caps lock is on and shift is held, the wrong case key will be used.
This is hard to fix since caps lock is OS dependent. I choose to ignore
it because it's hard and almost no one will care.

Related #153