src/gui/mpv/mpvwidget: revert dbus path for org.freedesktop.screensaver
This reverts commit f653de61129626b2d350146f6b2b61a8625b2c00.

On both Plasma and GNOME the path is /org/freedesktop/ScreenSaver. On
Flatpak it is just /ScreenSaver. This commit broke screen dimming for
normal users, so it's getting reverted. From now on Flatpaks will just
be patched to change the path and the main repo will use the standard
path.

I hate DBus