src/gui/widgets/mpv/mpvwidget: disable hwdec by default
hwdec=auto-safe was enabled by default because I thought it would be
desirable to use hardware decoding. Turns out that the warnings about
hwdec were not exagerated and it can cause problems in some not
unreasonable cases. Since software decoding is always the safer option,
I'll be leaving it to "power users" to enable the feature in their own
mpv.conf rather than risking creating a bad experience for people by
default.

Closes #144