src/main: set qtextcodec for locale to utf-8
On Windows QTextStream assumes the default code page for files without
a byte-order mark. This is much of UTF-8 since it isn't required and
sometimes isn't event recommended. On macOS and Linux this will almost
never be an issue because they use UTF-8 as the default code page.
Windows uses 1252 if you're an English speaker, if you're not then lol
good luck.

May God judge Bill Gates harshly for his crimes against unicode.

Closes #178