src/anki/ankiclient: send files using base64 over json
Use base64 encoding to send files to Anki. This has the benefit of
bypassing filesystem restrictions when using local files. It comes at
the cost of sending 33% more data. For the small file sizes being delt
with, this is unimportant. This also fixes the rare case where the
AnkiConnect server is on a different computer than Memento.

Closes #189