src/gui/widgets/mpvwidget: fix crash when hide cursor with osc is unset
If 'Hide Cursor with OSC' is unset, it causes Memento to crash on
launch. This happens because MpvTimer interacts with the mpv_handle
before it is fully initialized. This causes the code to use strcmp on
a null pointer. This moves the call to initTimer() into initializeGL()
which fixes the issue.

Fixes #192
Maybe Fixes #172