src/player/mpvadapter: fix handling of some numpad keys
mpv has a wacko way for dealing with keypresses that's filled with
little pit falls. Essentially the KP modifier is not constent with
number keys having no underscore (KP1), actions having an underscore
(KP_ENTER), and arethmetic being handled as if it were a normal key.
This is not documented anywhere except source code. This fixes the issue
by handing the keys modified by the numpad modifier on a case by case
basis.

Fixes #195