cmakelists: migrate to qt6
Migrates Memento to Qt6. This is important since Plasma 6 is out of beta
in February. There will likely not be a new release until then. Reworks
SubtitleListWidget to use std::shared_ptr since what was safe in Qt5
using QList is no longer safe in Qt6 due to the merging of the QList and
QVector implementations.