src/dict: abstract out deconjugators from dictionary code
Create a new class called QueryGenerator that defines a class that can
generate queries from a string of text. ExactQueryGenerator implements
this for searching exact text, and MeCabQueryGenerator implements this
for deconjuating text with MeCab. This makes it so these two methods of
generating queries can share all the same code for filtering duplicates,
generating cloze info, and searching the database. This also makes it
easier to implement new query generators in the future such as the
deconjuator in #210 or the proposed MeCab with UniDic option in #211.

This commit removes all multithreading from searches. I have known for
a while that it provides little to no performance imporvement due to
most time being spent querying a synchronized database. It may also harm
performance on Windows. I don't see any need to add it back, but it can
be if there is a compelling reason to. Queries seem plenty fast with
this change.