src/player/mpvadapter: fix files not playing on mpv v0.38.0
Something changed so mpv v0.38.0 won't accept loadfile commands if the
option string is an empty string. This is annoying because it completely
breaks Memento with relatively little warning since this was a change
made in the last couple weeks before release. This fixes the issue by
checking if the option string is empty and using NULL in that case.