src/gui/widgets/settings/searchsettings: add ability to toggle matchers
Add ability to select which matchers are enabled to support searching.
This allows the ExactQueryGenerator, MeCabQueryGenerator, and
DeconjugationQueryGenerator to be disabled or enabled independently of
each other based on user preference. This is included to support users
who only want to use MeCab as their query generator.

This also makes a minor change to term result sorting so that terms with
deconjugation information are prioritized over terms without. This is
because MeCab deconjugations tend to beat out deconjugator conjugations
when both are enabled. I find this to be undesirable.