src/gui/widgets/overlay/playermenu: fix unreadable text in dark mode
On Windows if the user enabled dark mode, the text would be unreadable.
This is due to a Qt bug as far as I can tell. This makes it so the Alt
proxy style isn't used on Windows when dark mode is set. macOS and Linux
remain unaffected by this change.