src/dict/dictionary: fix crash due to strict weak ordering in std::sort
The check for conjugation explination in sortTerm's first call to
std::sort creates scenario where strict weak ordering is possible. If
this condition is ever encountered, std::sort segfaults. This is my
fault as I wrote this function. Sorting based on conjugationExplanation
is changed to only occur if one explanation is empty and the other is
non-empty. This seems to fix the issue.

Closes #229