src/anki: add infrastructure for handling extended marker syntax
Add infrastructure to handle extended marker syntax. This one rather
than just replying on find and replace, first each field is parsed by
something that picks out tokens and organizes them into something that
can be handled easily by software. The marker is then matched to a known
marker and a handler is used. As of right now, this just replaces the
raw marker match with the proper content.

When extended markers are used, more intricate handlers will be
available to handle replacing markers. Using the raw value to replace
the text will allow markers of the same type with different arguments to
exist in the same field. For example { glossary : dic-id = 1 } and
{ glossary : dic-id = 3 } can co-exist. This is inefficient in cases
where markers are subtly different but use the same arguments, but those
are so unlikely, they aren't worth optimizing for.