src/gui/widgets/subtitlelistwidget: prevent stealing shortcuts
Prevent the subtitle list from stealing shortcut presses when it doesn't
have focus. Previously if the subtitle list was open, it would prevent
copying from other parts of the UI with CTRL+C. This could be annoying
when copying from glossaries. This fixes the issue by setting the
QShortcut context to Qt::WidgetWithChildrenShortcut and chaning the
focus of the subtitle list to Qt::ClickFocus. This does have the
downside of requiring the subtitle list be clicked before copying lines,
but that makes sense from a design perspective.

Closes #261