src/gui/widgets/subtitlelistwidget: add options to list find
Add ability to ignore whitespaces in both query text and the subtitles
being searched. Everything will be treated as if there is no whitespace.
Add the ability to automatically seek to found subtitles. Both of these
options exist as checkboxes above the search input. This also fixes some
bad tooltips on the find next and find previous buttons.

Closes #264