src/gui/widgets/overlay/playermenu: fix bad theming on windows
The menubar would have a bad style on launch that would lead to
unreadable text when using dark mode. This fixes the issue by removing
the initTheme() calls from PlayerMenu. It seems like this code exists to
fix a bug in Qt5. Removing code is always a welcome change.

Related #260