src/anki/notebuilder: implement coditional marker syntax
Allow markers to use conditional syntax with the | character. This makes
it so if a marker has empty text, the next marker will be used to
replace it. This can be useful in cases such as {selection | clipboard}
where if nothing is selected, the clipboard can be used instead. Media
markers such as {audio} will not trigger the next marker to be used
since they are not considered empty.

Related #262