src/anki/ankiclient: refactor to use coroutines for ankiconnect calls
Refactor to use coroutines with QCoro. This greatly simplifies the code
as each additional request to Anki doesn't need to be nested in
connect() calls. This will also benefit the code anywhere that network
requests will be used. That will come in a later PR.

Adds QCoro v0.12.0 to FetchContent in extern/CMakeLists.txt