src/gui/widgets/definition/termwidget: fix crash on add
Fixes a crash when adding a term. This is due to a race where the
TermWidget is deleted before addNote finally ends execution. This used
to not be an issue before the coroutine refactor, but was reintrouced
since m_ankiTerm no longer uses a raw pointer. This replaces the check
with a boolean instead.