src/gui/widget/overlay/subtitlewidget: allow scan using middle mouse
This feature adds the option to scan terms using middle mouse button.
A new option in the setting page was added to handle middle mouse scan.
Since this feature only makes sense when search method is not hover, the
new checkbox and the modifier settings are now bundled under a layout,
and the layout will be visible only when search method is equal to
modifier (similar to scan delay being only visible when search method
is hover).