src/anki/notebuilder: allow filtering only lowest frequency
This feature extends the ``{frequencies}`` marker with the extended
marker syntax and adds two new options: ``min-value`` which allows to
export only the lowest value frequency out of all the options and
``value-only`` which strips the dictionary name and html list format to
keep only raw numbers.

This allow memento to support card format that rely on raw frequency
values to sort their cards.