src/gui/widgets/definition/termwidget: make term widget safe to delete
For the longest time TermWidget was a disaster for memory safety in
this program because it relied on network requests in several places
that could return after their parent TermWidget had already been
deleted. Using Qt's signals and slots, this would lead to cases where
the network request completes and since TermWidget is gone, the rest of
the code that handles that finished signal no longer executes. This
could lead to things like cards not getting added to Anki which is
hopelessly annoying. For the longest time, TermWidget had a bonkers
implementation where a safeToDelete() function was used to signal that
the TermWidget should be deleted after it's finished doing whatever it
was doing. This required dynamic_casts as was just plain terrible. This
new implementation removes all of that.

By leveraging static methods and using QCoro::Task, all the asynchronous
methods can complete safely even if TermWidget gets deleted. Using
QCoro::connect, custom callbacks can be handled that aren't unsafe since
they don't execute if TermWidget gets deleted. The code quality becomes
so much nicer as result.