src/gui/widgets/definition/kanjiwidget: improve safety of anki functions
Make add and opening in Anki safer. Previously it was possible but
highly unlikely for dangling KanjiWidget references to occur. This fixes
the issue by making all the coroutines static and only connecting
non-static functions via QCoro::connect. This is similar to the fixes
that have happened in TermWidget but much simpler due to the lack of
chained coroutines.