src/anki/notebuilder: add support for cliping videos
Add support for cliping videos with the {video} marker. This supports
the extended marker syntax:
 * source - Specifies the source of the video timings
   * subtitle - The current subtitle is used to generate the timing
   * context - The current context is used to generate the timing
 * audio - True if audio should be included, false otherwise
 * subtitles - True if subtitles should be included, false otherwise
 * norm - True if audio should be normalized, false otherwise
 * db - The decibels to normalize audio to if norm = true
Currently this is undocumented in the program. I'd like to make the
documentation a little easier to maintain, so I'll leave documenting it
for after that is done.

Makes tempAudioClip and tempVideoClip take a struct for arguments since
that makes it easier to implement optional arguments.

Closes #294