option/cmakelists: add option to usage system installation of qcoro
Probably useful for teto's Nix stuff. Will also be useful for the AUR.
Disabled by default.

This implementation of this is ugly because QCoro has different include
paths when installed verses when linked to directly. It also has
different CMake targets. This slings together the ugliest generator
expressions possible to handle this and uses ifdefs for every QCoro
include. I hate it.